pyproject.toml
[tool.poetry]
name = "twarm"
version = "0.1.0"
description = ""
authors = ["Your Name <OverflowCat@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Telethon = "^1.21.1"
python-dotenv = "^0.15.0"
tinydb = "^4.4.0"
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"